pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "scroogebot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.0",
    "yfinance>=0.2",
    "ta>=0.11",
    "sqlalchemy>=2.0",
    "aiomysql>=0.2",
    "pymysql>=1.1",
    "alembic>=1.13",
    "apscheduler>=3.10",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov",
]
backtest = [
    "vectorbt>=0.26",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
